Web Scraping/requirements.txt
bs4==0.0.1
requests==2.28.2
lxml==4.9.2
Pillow==9.5.0